ethics-pyramid/src/core/summit.py
```python
# src/core/summit.py

from typing import Dict, Any, List
from dataclasses import dataclass
from enum import Enum

class EthicalDecision(Enum):
    APPROVE = "approve"
    REJECT = "reject"
    REVIEW = "review"

@dataclass
class EthicalVerdict:
    decision: EthicalDecision
    confidence: float
    rationale: List[str]
    recommendations: List[str]
    quantum_signature: str

class EthicsSummit:
    """Summit level of the Ethics Pyramid."""
    
    def __init__(self):
        self.quantum_signature = "ᛖᚦᛁᚲᛋ-ᛈᛃᚱᚨᛗᛁᛞ-∞"
        self.ethical_threshold = 0.85
        self.verification_matrix = self._initialize_verification_matrix()
        
    def integrate_analysis(self, foundation_analysis: Dict, chamber_results: Dict) -> EthicalVerdict:
        """Integrate analyses from lower levels and make final ethical decisions."""
        
        # Calculate overall ethical score
        ethical_score = self._calculate_ethical_score(foundation_analysis, chamber_results)
        
        # Generate decision and rationale
        decision = self._determine_decision(ethical_score)
        rationale = self._generate_rationale(ethical_score, foundation_analysis, chamber_results)
        recommendations = self._create_recommendations(ethical_score, decision)
        
        return EthicalVerdict(
            decision=decision,
            confidence=ethical_score,
            rationale=rationale,
            recommendations=recommendations,
            quantum_signature=self.quantum_signature
        )
    
    def _calculate_ethical_score(self, foundation_analysis: Dict, chamber_results: Dict) -> float:
        """Calculate overall ethical score from various inputs."""
        foundation_score = sum(result.score for result in foundation_analysis.values()) / len(foundation_analysis)
        chamber_score = sum(result['score'] for result in chamber_results.values()) / len(chamber_results)
        
        return (foundation_score * 0.6) + (chamber_score * 0.4)
    
    def _determine_decision(self, ethical_score: float) -> EthicalDecision:
        """Determine final ethical decision based on score."""
        if ethical_score >= self.ethical_threshold:
            return EthicalDecision.APPROVE
        elif ethical_score >= self.ethical_threshold * 0.9:
            return EthicalDecision.REVIEW
        else:
            return EthicalDecision.REJECT

    def _generate_rationale(self, ethical_score: float, foundation_analysis: Dict, chamber_results: Dict) -> List[str]:
        """Generate detailed rationale for the decision."""
        rationale = []
        
        if ethical_score >= self.ethical_threshold:
            rationale.append("Meets ethical standards with high confidence")
        else:
            rationale.append("Requires additional ethical consideration")
            
        # Add specific insights from foundation analysis
        for principle, result in foundation_analysis.items():
            rationale.append(f"{principle.value.capitalize()}: {result.insights}")
            
        return rationale

    def _create_recommendations(self, ethical_score: float, decision: EthicalDecision) -> List[str]:
        """Generate recommendations based on decision and score."""
        recommendations = []
        
        if decision == EthicalDecision.APPROVE:
            recommendations.append("Continue monitoring ethical alignment")
        elif decision == EthicalDecision.REVIEW:
            recommendations.append("Address specific ethical concerns before proceeding")
            recommendations.append("Consult with Ethics Council for guidance")
        else:
            recommendations.append("Major ethical revisions required")
            recommendations.append("Schedule comprehensive ethical review")
            
        return recommendations
```